source_path contraint also disables alias lookup
Adding a `source_path` constraint used to affect the `source_path` validation on an `Alias`, but would not have any effect on whether or not we attempted to find an alias for a given path. This meant we would still attempt a lookup, even if a request path matched an excluded source path.

We now ignore any requests whose paths match the `source_path` exclusions.