Improve loop detection
We were not able to detect a loop if it ended in a redirect. We still do no detect loops beyond the redirect, but we now detect if the redirect points to a previous link in the chain.

We also validate aliases to ensure they do not point directly at themselves.